+++ modules/statistics/app/controllers/statistics/Statistics.scala
@@ -2,7 +2,7 @@ package controllers.statistics
 
 import controllers.OrganizationController
 import play.api.mvc.Action
-import models.{DomainConfiguration, DataSet}
+import models.{Role, Group, DomainConfiguration, DataSet}
 import collection.JavaConverters._
 import models.statistics.DataSetStatistics
 import collection.immutable.ListMap
@@ -76,8 +76,7 @@ object Statistics extends OrganizationController {
    *  The purpose of this fu
    */
 
-  def legacyStatistics(orgId: String) = {
-    DomainConfigured {
+  def legacyStatistics(orgId: String) = DomainConfigured {
       Action {
         implicit request =>
 
@@ -92,11 +91,19 @@ object Statistics extends OrganizationController {
             }
           }
 
-          val statistics = new SolrFacetBasedStatistics(facets, orgId)
-          Ok(statistics.renderAsJSON()).as(JSON)
+          val filter = request.queryString.get("filter").flatMap { f =>
+            f.headOption
+          }
+
+          val canSeeFullStatistics = request.session.get(Constants.USERNAME).map { userName =>
+            Group.dao.hasRole(userName, StatisticsPlugin.UNIT_ROLE_STATISTICS_VIEW) || Group.dao.hasRole(userName, Role.OWN)
+          }.getOrElse(false)
+
+          val statistics = new SolrFacetBasedStatistics(orgId, facets, filter)
+          Ok(statistics.renderAsJSON(canSeeFullStatistics)).as(JSON)
       }
     }
-  }
+
 }
 
 case class StatisticsCounter(name: String, total: Int, withNr: Int = 0)  {
@@ -138,7 +145,7 @@ case class StatisticsHeader(name: String, label: String = "", entries: Seq[Combi
   }
 }
 
-class SolrFacetBasedStatistics(facets: Map[String, String], orgId: String)(implicit configuration: DomainConfiguration, lang: Lang) {
+class SolrFacetBasedStatistics(orgId: String, facets: Map[String, String], filter: Option[String])(implicit configuration: DomainConfiguration, lang: Lang) {
 
     val orgIdFilter = "%s:%s".format(IndexField.ORG_ID.key, orgId)
 
@@ -151,6 +158,7 @@ class SolrFacetBasedStatistics(facets: Map[String, String], orgId: String)(impli
     query addFacetField (facetsForStatistics: _ *)
     query setRows (0)
     query setFilterQueries (orgIdFilter)
+    filter foreach { f => query addFilterQuery f }
 
     val allRecordsResponse = SolrQueryService.getSolrResponseFromServer(solrQuery = query)
     val allRecords = SolrBindingService.createFacetStatistics(allRecordsResponse.getFacetFields.asScala.toList)
@@ -158,12 +166,14 @@ class SolrFacetBasedStatistics(facets: Map[String, String], orgId: String)(impli
 
     // query for with only digital objects
     query setFilterQueries("%s:true".format(IndexField.HAS_DIGITAL_OBJECT.key), orgIdFilter)
+    filter foreach { f => query addFilterQuery f }
     val digitalObjectsResponse = SolrQueryService.getSolrResponseFromServer(solrQuery = query)
     val digitalObjects = SolrBindingService.createFacetStatistics(digitalObjectsResponse.getFacetFields.asScala.toList)
     val totalDigitalObjects = digitalObjectsResponse.getResults.getNumFound
 
     // query with landing pages
     query setFilterQueries("%s:[* TO *]".format(SystemField.LANDING_PAGE.tag), orgIdFilter)
+    filter foreach { f => query addFilterQuery f }
     val landingPagesResponse = SolrQueryService.getSolrResponseFromServer(solrQuery = query)
     val landingPages = SolrBindingService.createFacetStatistics(landingPagesResponse.getFacetFields.asScala.toList)
     val totalLandingPages = landingPagesResponse.getResults.getNumFound
@@ -200,18 +210,28 @@ class SolrFacetBasedStatistics(facets: Map[String, String], orgId: String)(impli
 
   val entries: Seq[StatisticsHeader] = facets.map(createHeader(_)).toSeq
 
-  def renderAsJSON(): String = {
+  val entryCounts: Map[String, Int] = entries.map { e =>
+    (e.name -> e.entries.size)
+  }.toMap
+
+  def renderAsJSON(displayFacetDetail: Boolean): String = {
     import net.liftweb.json.{Extraction, JsonAST, Printer}
     implicit val formats = net.liftweb.json.DefaultFormats
 
     val outputJson = Printer.pretty(JsonAST.render(Extraction.decompose(
-      ListMap("statistics" ->
-        ListMap(
+      ListMap("statistics" -> {
+        val stats = ListMap(
           "totalRecords" -> totalRecords,
           "totalRecordsWithDigitalObjects" -> totalDigitalObjects,
           "totalRecordsWithLandingPages" -> totalLandingPages,
-          "facets" -> entries.map(_.asListMap)
+          "facetCounts" -> entryCounts
         )
+        if (displayFacetDetail) {
+          stats ++ ListMap("facets" -> entries.map(_.asListMap))
+        } else {
+          stats
+        }
+      }
       ))))
 
     outputJson